+++ src/main/java/mhrda/socialme/utilities/Size.java
@@ -2,26 +2,26 @@ package mhrda.socialme.utilities;
 
 public class Size {
 	
-	private int width;
-	private int height;
+	private double width;
+	private double height;
 	
 	public Size() {}
 	
-	public Size(int width, int height) {
+	public Size(double width, double height) {
 		this.width = width;
 		this.height = height;
 	}
 	
-	public int getWidth() {
+	public double getWidth() {
 		return width;
 	}
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
-	public int getHeight() {
+	public double getHeight() {
 		return height;
 	}
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
